cleanup comments, move gamestate
