+++ game/src/gamestate.rs
@@ -1,0 +1,41 @@
+use crate::prelude::*;
+
+pub struct GameStatePlugin;
+
+impl Plugin for GameStatePlugin {
+    fn build(&self, app: &mut App) {
+        app.add_state::<GameState>();
+    }
+}
+
+#[derive(States, Clone, Eq, PartialEq, Hash, Debug, Default)]
+pub enum GameState {
+    #[default]
+    Playing,
+    Paused,
+}
+
+pub fn pause(
+    mut next_state: ResMut<NextState<GameState>>,
+    input: Res<Input<KeyCode>>,
+    mut time: ResMut<Time<Physics>>,
+) {
+    if input.just_pressed(KeyCode::P) {
+        next_state.set(GameState::Paused);
+        //pause physics
+        time.pause();
+    }
+}
+
+pub fn unpause(
+    mut next_state: ResMut<NextState<GameState>>,
+    input: Res<Input<KeyCode>>,
+    mut time: ResMut<Time<Physics>>,
+) {
+    if input.just_pressed(KeyCode::P) {
+        next_state.set(GameState::Playing);
+        time.unpause();
+    }
+}
+
+
